text_recognition/requirements.txt
vietocr
matplotlib
pyyaml
torch
torchvision
pillow
# opencv-python
einops==0.2.0
gdown==4.4.0
prefetch_generator==1.0.1
imgaug==0.4.00
lmdb==1.0.0